Title Get ERDDAP gridded data restricted to a given season of the year

griddap_season uses the R program 'rerddap::griddap()' to extract environmental data from an 'ERDDAP' server in an (time, z, y ,x) bounding box where time is restricted to a given season of the year (see below). Arguments are the same as in 'rerddap::griddap()' except for the added 'season' parameter. 'read' and 'fmt' options are ignored.

griddap_split(
  datasetx,
  ...,
  fields = "all",
  stride = 1,
  request_split = NULL,
  fmt = "nc",
  url = rerddap::eurl(),
  store = rerddap::disk(),
  read = TRUE,
  callopts = list(),
  aggregate_file = NULL
)

Arguments

datasetx

Anything coercable to an object of class info. So the output of a call to info, or a datasetid, which will internally be passed through info

...

Dimension arguments. See examples. Can be any 1 or more of the dimensions for the particular dataset - and the dimensions vary by dataset. For each dimension, pass in a vector of length two, with min and max value desired. at least 1 required.

fields

(character) Fields to return, in a character vector.

stride

(integer) How many values to get. 1 = get every value, 2 = get every other value, etc. Default: 1 (i.e., get every value)

request_split

A numeric vector indicating the number of splits for each dimension, used to segment the request into manageable chunks. This is particularly useful for large datasets.

fmt

(character) One of:

nc: save output to a netcdf file

memory: save output in a dataframe in memory

duckdb: save data in a duckdb database

url

A URL for an ERDDAP server. Default: https://upwell.pfeg.noaa.gov/erddap/ - See rerddap::eurl() for more information

store

ignored

read

ignored

callopts

Curl options passed on to verb-GET

aggregate_file

A string specifying:

if is.null(aggregate_file) then a temporary file is created that does not overwrite user space

if "fmt = nc" the path to the NetCDF file to store the results

if "fmt = duckdb" the path to the duckdb file to store the results

if "fmt = memory" the value is ignored

Value

Varies by format requested:

if "fmt = nc" the path to the NetCDF file where the results are stored

if "fmt = duckdb" the path to the duckdb file where the results are stored

if "fmt = memory" the usual rerddap::griddap dataframe

Examples

out <- rerddap::info('erdQMekm14day')
request_split <- list(time = 3, altitude = 1, latitude = 1, longitude = 1)
res <- griddap_split(out,
                     time = c('2015-12-28','2016-01-01'),
                     latitude = c(20, 40),
                     longitude = c(220, 240),
                     fields = 'mod_current',
                     request_split = request_split
                     )
#> info() output passed to x; setting base url to: https://upwell.pfeg.noaa.gov/erddap
#> info() output passed to x; setting base url to: https://upwell.pfeg.noaa.gov/erddap